Add files via upload
The goal is to compute the number of businesses, total review count, and number of 4-star or higher reviews for each neighborhood in each city. If a business has multiple neighborhoods, its review count and stars should be attributed to all of the neighborhoods. If the neighborhoods list is empty, then we will use 'Unknown' as the name of the neighborhood.